# Add a couple of indexes to speed up some queries