```
export-service=# explain analyze SELECT count(*) FROM "export_payloads" WHERE "export_payloads"."account_id" = '1' AND "export_payloads"."organization_id" = '2' AND "export_payloads"."username" = 'fred';                                                   QUERY PLAN                                                   
----------------------------------------------------------------------------------------------------------------
 Aggregate  (cost=47.52..47.52 rows=1 width=8) (actual time=0.231..0.231 rows=1 loops=1)
   ->  Seq Scan on export_payloads  (cost=0.00..47.51 rows=1 width=0) (actual time=0.229..0.229 rows=0 loops=1)
         Filter: ((account_id = '1'::text) AND (organization_id = '2'::text) AND (username = 'fred'::text))
         Rows Removed by Filter: 1115
 Planning Time: 0.099 ms
 Execution Time: 0.249 ms
(6 rows)

export-service=# CREATE INDEX CONCURRENTLY IF NOT EXISTS export_payloads_account_id_org_id_username_index ON export_payloads (account_id, organization_id, username);
CREATE INDEX
export-service=# explain analyze SELECT count(*) FROM "export_payloads" WHERE "export_payloads"."account_id" = '1' AND "export_payloads"."organization_id" = '2' AND "export_payloads"."username" = 'fred';
                                                                                 QUERY PLAN        
                                                                          
---------------------------------------------------------------------------------------------------
--------------------------------------------------------------------------
 Aggregate  (cost=6.05..6.06 rows=1 width=8) (actual time=0.011..0.011 rows=1 loops=1)
   ->  Index Only Scan using export_payloads_account_id_org_id_username_index on export_payloads  (
cost=0.28..6.05 rows=1 width=0) (actual time=0.008..0.008 rows=0 loops=1)
         Index Cond: ((account_id = '1'::text) AND (organization_id = '2'::text) AND (username = 'fred'::text))
         Heap Fetches: 0
 Planning Time: 0.136 ms
 Execution Time: 0.029 ms
```