@dehort 2024-09-12
DB migration up and down seems to work